+++ src/errors.rs
@@ -4,6 +4,8 @@
 //! https://github.com/DanielKeep/rust-error-type/issues/2.
 
 use redis::RedisError;
+use std::io;
+use std::borrow::Cow;
 
 
 /// A ``SpaceapiServerError`` wraps general problems that can occur in the SpaceAPI server.
@@ -12,6 +14,14 @@ error_type! {
     pub enum SpaceapiServerError {
         Redis(RedisError) {
             cause;
-        }
+        },
+        IoError(io::Error) {
+            cause;
+        },
+        Message(Cow<'static, str>) {
+            desc (e) &**e;
+            from (s: &'static str) s.into();
+            from (s: String) s.into();
+        },
     }
 }