+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Possible log types:
 - [changed] Removed datastore module, use Redis directly (#10)
 - [changed] SpaceapiServer.serve() now returns a HttpResult<Listening> (#16)
 - [added] Support status modifiers (#8)
+- [changed] Use `ToSocketAddrs` instead of `IPv4addr` and port in `SpaceapiServer::new()` (#22)
 
 ### v0.1.1 (2015-11-16)